requirements.txt
# PlugVerse - Core Dependencies
# ================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Async Support
aiofiles>=23.0.0
anyio>=4.0.0

# Logging
loguru>=0.7.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Database (Optional)
sqlalchemy>=2.0.0

# Redis (Optional, for task queue)
# redis>=5.0.0
# celery>=5.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
